requirements.txt
pycryptodome==3.21.0
PyYAML==6.0.2
